// backend-admin-site/src/core/kernel/presentation/presentation.types.ts
export type PresentationDeviceType =
  | "mobile"
  | "tablet"
  | "web"
  | "desktop"
  | "laptop"
  | "tv"
  | "monitor"
  | "projector";

export type PresentationTransportType =
  | "wifi"
  | "webrtc"
  | "chromecast"
  | "miracast"
  | "airplay"
  | "dlna"
  | "hdmi";

export type PresentationSourceType =
  | "screen"
  | "app"
  | "video_call"
  | "camera"
  | "photo"
  | "file"
  | "browser_tab";